Render doctext citations in Chat too
Doctext links — the citations a reply uses to point at the writer's own
text, which select that text when clicked — were Revise-only, built
inline in its page component. Chat asks about the same document and has
the same reason to say "look here", so the machinery moves to
`components/docTextLink/` and both pages use it.

The shared module owns the whole behaviour, not just the anchor: the jump
(with its sequencing, so a slow search can't clear a newer click's
indicator), the pending spinner, the not-found note, the screen-reader
status region, and the `urlTransform` that exempts `doctext:` from React
Markdown's URL filter. A page needs three things now — `useDocJump(page)`,
`<DocTextMarkdown jump={…}>` in place of `<Markdown>`, and one
`<DocJumpStatus>` — which is what made adding the second page small.
`docTextJump.ts` and its tests move alongside, unchanged.

Chat also needs to *ask* for the links, since nothing emits them
otherwise: `CHAT_INSTRUCTIONS` now carries a short form of the rules
Revise's prompt spells out (encoded verbatim quote, single line, 240
characters, description as the link text).

Logging: `reference_clicked` / `reference_resolved` were Revise-scoped
helpers. They become `referenceLog`, which takes `page` as a parameter —
the shape `briefLog` already uses for behaviour that isn't one page's.
`LOG_SCHEMA_VERSION` goes to 5, because a reader that took those event
names to mean Revise now has to read `page` instead.

Verified end to end against a production build: a citation in a chat
reply keeps its `doctext:` href through the URL filter and selects the
quoted text on click, and a citation quoting text that isn't in the
document shows the not-found affordance.
