+++ frontend/src/pages/revise/styles.module.css
@@ -385,66 +385,5 @@
 	}
 }
 
-/*
- * Doctext links: clicking one asks the editor to find and select the quoted
- * text, which is a round-trip on the Google Docs surface. The link itself
- * carries the progress — a link that looks unchanged for a second reads as a
- * broken link, and the writer clicks again.
- */
-.docLink {
-	cursor: pointer;
-}
-
-.docLinkPending {
-	cursor: progress;
-	opacity: 0.75;
-}
-
-.linkSpinner {
-	display: inline-block;
-	width: 9px;
-	height: 9px;
-	margin-left: 5px;
-	vertical-align: baseline;
-	border: 1.5px solid currentColor;
-	border-top-color: transparent;
-	border-radius: 50%;
-	animation: linkSpin 0.7s linear infinite;
-}
-
-.linkFailed {
-	margin-left: 6px;
-	font-size: 11px;
-	font-style: italic;
-	color: var(--text-secondary);
-	white-space: nowrap;
-}
-
-@keyframes linkSpin {
-	to {
-		transform: rotate(360deg);
-	}
-}
-
-/* Announced by screen readers, invisible on screen. */
-.visuallyHidden {
-	position: absolute;
-	width: 1px;
-	height: 1px;
-	margin: -1px;
-	padding: 0;
-	overflow: hidden;
-	clip: rect(0 0 0 0);
-	clip-path: inset(50%);
-	white-space: nowrap;
-	border: 0;
-}
-
-/* Respect a reduced-motion preference: keep the state, drop the animation. */
-@media (prefers-reduced-motion: reduce) {
-	.linkSpinner {
-		animation: none;
-		border-top-color: currentColor;
-		opacity: 0.5;
-	}
-}
+/* Doctext link and jump-status styles live in `components/docTextLink/`,
+   shared with Chat. */
